requirements.txt
Pillow==10.2.0
pytesseract==0.3.10
icalendar==5.0.11
python-dotenv==1.0.0 